Add milliseconds-based version of formatDateOrTime() and isThisYear(), deprecated seconds-based version
The seconds-based version uses Int so it will break due to the Year
2038 problem. Add a version that uses milliseconds (as Long).

This matches the existing formatDate() extension, which already had
both a seconds-based (Int) version and a milliseconds-based (Long)
version.

Also deprecate Int.formatDate(), and make it a simple wrapper for
Long.formatDate() to reduce code repetition.
