+++ commons/src/main/kotlin/org/fossify/commons/extensions/Long.kt
@@ -2,6 +2,8 @@ package org.fossify.commons.extensions
 
 import android.content.Context
 import android.text.format.DateFormat
+import android.text.format.DateUtils
+import android.text.format.Time
 import java.text.DecimalFormat
 import java.util.Calendar
 import java.util.Locale
@@ -22,4 +24,35 @@ fun Long.formatDate(context: Context, dateFormat: String? = null, timeFormat: St
     val cal = Calendar.getInstance(Locale.ENGLISH)
     cal.timeInMillis = this
     return DateFormat.format("$useDateFormat, $useTimeFormat", cal).toString()
+}
+
+// if the given date is today, we show only the time. Else we show the date and optionally the time too
+fun Long.formatDateOrTime(context: Context, hideTimeAtOtherDays: Boolean, showYearEvenIfCurrent: Boolean): String {
+    val cal = Calendar.getInstance(Locale.ENGLISH)
+    cal.timeInMillis = this
+
+    return if (DateUtils.isToday(this)) {
+        DateFormat.format(context.getTimeFormat(), cal).toString()
+    } else {
+        var format = context.baseConfig.dateFormat
+        if (!showYearEvenIfCurrent && isThisYear()) {
+            format = format.replace("y", "").trim().trim('-').trim('.').trim('/')
+        }
+
+        if (!hideTimeAtOtherDays) {
+            format += ", ${context.getTimeFormat()}"
+        }
+
+        DateFormat.format(format, cal).toString()
+    }
+}
+
+fun Long.isThisYear(): Boolean {
+    val time = Time()
+    time.set(this)
+
+    val thenYear = time.year
+    time.set(System.currentTimeMillis())
+
+    return (thenYear == time.year)
 }
